# Provide concrete-only traversal for Class#subclasses