For classes with many singletons, the Class#subclasses walk will cost significantly more than is warranted for the few concrete classes it will return. This can be eliminated by tracking whether each entry is concrete and adding a separate traversal link to the next concrete class in the chain.

Performance of such a case is significantly better:

Before:
```
1 thread Numeric.subclasses
                          1.315k (± 4.6%) i/s -      6.612k in   5.040800s
1 thread Object.subclasses
                         56.282 (± 3.6%) i/s -    285.000 in   5.074688s
1 thread Custom.subclasses with 1000 singletons
                          5.588 (± 0.0%) i/s -     28.000 in   5.021174s
```

After:
```
1 thread Numeric.subclasses
                          2.402k (± 3.4%) i/s -     11.990k in   4.998757s
1 thread Object.subclasses
                        189.977 (± 6.8%) i/s -    954.000 in   5.068134s
1 thread Custom.subclasses with many singletons
                          3.439k (± 1.0%) i/s -     17.334k in   5.040639s
```